[onert/lib] Move RandomTestRunner class
Introduce RandomTestRunner.cpp/h files and move RandomTestRunner class from Diff.cpp/h
