+++ runtime/libs/tflite/include/tflite/Diff.h
@@ -85,64 +85,4 @@ private:
   const nnfw::misc::tensor::Comparator &_comparator;
 };
 
-#include "tflite/interp/Builder.h"
-#include "tflite/Quantization.h"
-
-/**
- * @brief Structure for NNAPI correctness test
- */
-struct RandomTestParam
-{
-  int verbose;               //!< Verbosity of debug information
-  int tolerance;             //!< Torlerance of value difference
-  int tensor_logging = 0;    //!< Save logging to a file if not 0
-  std::string log_path = ""; //!< Path of log file, meaningful only when tensor_logging is 1
-};
-
-/**
- * @brief Class to define Random test runner
- */
-class RandomTestRunner
-{
-public:
-  /**
-   * @brief Construct a new RandomTestRunner object
-   * @param[in] seed          Random seed value
-   * @param[in] param         RandomTestParam object for test runner
-   * @param[in] quantization  TfLiteQuantizationParams type to represent quantization value
-   */
-  RandomTestRunner(uint32_t seed, const RandomTestParam &param)
-      : _randgen{seed, 0.0f, 2.0f}, _param{param}
-  {
-    // DO NOTHING
-  }
-
-public:
-  /**
-   * @brief  Run the random test runner
-   * @param[in] builder  Interpreter Builder used to run
-   * @return 0 if test succeeds, otherwise failure
-   */
-  int run(const nnfw::tflite::Builder &builder);
-
-public:
-  /**
-   * @brief  Get RandomGenerator reference
-   * @return RandomGenerator reference
-   */
-  nnfw::misc::RandomGenerator &generator() { return _randgen; };
-
-private:
-  nnfw::misc::RandomGenerator _randgen;
-  const RandomTestParam _param;
-
-public:
-  /**
-   * @brief  Create a RandomTestRunner object
-   * @param[in] seed  Random seed value
-   * @return RandomGenerator object
-   */
-  static RandomTestRunner make(uint32_t seed);
-};
-
 #endif // __NNFW_TFLITE_DIFF_H__
